ugc/src/requirements.txt
aiokafka==0.8.0
anyio==3.6.2
cffi==1.15.1
click==8.1.3
clickhouse-driver==0.2.5
colorama==0.4.6
cryptography==39.0.1
fastapi==0.92.0
fastapi-jwt-auth==0.5.0
sniffio==1.3.0
starlette==0.25.0
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uvicorn==0.20.0
win32-setctime==1.1.0
gunicorn==20.1.0
confluent-kafka==1.9.2
pydantic[dotenv]
httpx==0.23.3